+++ lua/plugins/baseadmincommands.lua
@@ -34,19 +34,21 @@ local function PrintStatus(player, client, index)
 	if not playerClient then
 		Shared.Message("playerClient is nil in PrintStatus, alert Brian")
 	else
-		
+		local playerId = playerClient:GetUserId()
 		if DAK:GetClientCanRunCommand(client, "sv_status") then
 			local playerAddressString = IPAddressToString(Server.GetClientAddress(playerClient))
 			ServerAdminPrint(client, player:GetName() .. " : Game Id = " 
 			.. ToString(DAK:GetGameIdMatchingClient(playerClient))
-			.. " : Steam Id = " .. playerClient:GetUserId()
+			.. " : NS2 Id = " .. playerId
+			.. " : Steam Id = " .. GetReadableSteamId(playerId)
 			.. " : Team = " .. player:GetTeamNumber()
 			.. " : Address = " .. playerAddressString
 			.. " : Connection Time = " .. DAK:GetClientConnectionTime(playerClient))
 		else
 			ServerAdminPrint(client, player:GetName() .. " : Game Id = " 
 			.. ToString(DAK:GetGameIdMatchingClient(playerClient))
-			.. " : Steam Id = " .. playerClient:GetUserId()
+			.. " : NS2 Id = " .. playerId
+			.. " : Steam Id = " .. GetReadableSteamId(playerId)
 			.. " : Team = " .. player:GetTeamNumber())
 		end
 
@@ -540,24 +542,24 @@ local function OnCommandBanSelection(client, selectionnumber, page)
 	local targetclient = DAK.gameid[selectionnumber + (page * 8)]
 	if targetclient ~= nil then
 		local HeadingText = string.format("Please confirm you wish to ban %s?", DAK:GetClientUIDString(targetclient))
-		DAK:DisplayConfirmationMenuItem(GetSteamIdMatchingClient(client), HeadingText, Ban, nil, DAK:GetSteamIdMatchingClient(targetclient))
+		DAK:DisplayConfirmationMenuItem(DAK:GetNS2IdMatchingClient(client), HeadingText, Ban, nil, DAK:GetNS2IdMatchingClient(targetclient))
 	end
 end
 
 local function GetBansMenu(client)
-	DAK:CreateGUIMenuBase(DAK:GetSteamIdMatchingClient(client), OnCommandBanSelection, OnCommandUpdateBans, true)
+	DAK:CreateGUIMenuBase(DAK:GetNS2IdMatchingClient(client), OnCommandBanSelection, OnCommandUpdateBans, true)
 end
 
 local function GetKickMenu(client)
-	//DAK:CreateGUIMenuBase(DAK:GetSteamIdMatchingClient(client), OnCommandMenuVote, OnCommandUpdateVote, true)
+	//DAK:CreateGUIMenuBase(DAK:GetNS2IdMatchingClient(client), OnCommandMenuVote, OnCommandUpdateVote, true)
 end
 
 local function GetSlayMenu(client)
-	//DAK:CreateGUIMenuBase(DAK:GetSteamIdMatchingClient(client), OnCommandMenuVote, OnCommandUpdateVote, true)
+	//DAK:CreateGUIMenuBase(DAK:GetNS2IdMatchingClient(client), OnCommandMenuVote, OnCommandUpdateVote, true)
 end
 
 local function GetGagMenu(client)
-	//DAK:CreateGUIMenuBase(DAK:GetSteamIdMatchingClient(client), OnCommandMenuVote, OnCommandUpdateVote, true)
+	//DAK:CreateGUIMenuBase(DAK:GetNS2IdMatchingClient(client), OnCommandMenuVote, OnCommandUpdateVote, true)
 end
 
 DAK:RegisterMainMenuItem("Kick Menu", function(client) return DAK:GetClientCanRunCommand(client, "sv_kick") end, GetKickMenu)